+++ Unchase.OData.ConnectedService/ViewModels/AdvancedSettingsViewModel.cs
@@ -5,12 +5,16 @@ using System;
 using System.Threading.Tasks;
 using System.Windows;
 using Microsoft.VisualStudio.ConnectedServices;
+using Unchase.OData.ConnectedService.Common;
 using Unchase.OData.ConnectedService.Views;
 
 namespace Unchase.OData.ConnectedService.ViewModels
 {
     internal class AdvancedSettingsViewModel : ConnectedServiceWizardPage
     {
+        public Constants.FunctionImportsGenerator[] FunctionImportsGenerators =>
+            new[] { Constants.FunctionImportsGenerator.Inner, Constants.FunctionImportsGenerator.SimpleOData};
+
         public bool UseDataServiceCollection { get; set; }
         public bool UseNamespacePrefix { get; set; }
         public string NamespacePrefix { get; set; }
@@ -20,6 +24,7 @@ namespace Unchase.OData.ConnectedService.ViewModels
         public bool IncludeT4File { get; set; }
         public bool IncludeExtensionsT4File { get; set; }
         public Visibility IncludeExtensionsT4FileVisibility { get; set; }
+        public Constants.FunctionImportsGenerator FunctionImportsGenerator { get; set; }
 
         public AdvancedSettingsViewModel() : base()
         {
@@ -28,6 +33,7 @@ namespace Unchase.OData.ConnectedService.ViewModels
             this.Legend = "Settings";
             this.IncludeExtensionsT4File = false;
             this.IncludeExtensionsT4FileVisibility = Visibility.Collapsed;
+            this.FunctionImportsGenerator = Constants.FunctionImportsGenerator.Inner;
             this.ResetDataContext();
         }
 
@@ -59,6 +65,7 @@ namespace Unchase.OData.ConnectedService.ViewModels
             this.IncludeT4File = false;
             this.IncludeExtensionsT4File = false;
             this.IncludeExtensionsT4FileVisibility = Visibility.Collapsed;
+            this.FunctionImportsGenerator = Constants.FunctionImportsGenerator.Inner;
         }
     }
 }