+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 *Features:*
 
 - Generate client-side C# proxy-classes for OData protocol versions 1.0-4.0;
-- Generate client-side class with proxy-class extension methods for OData protocol versions 1.0-3.0;
+- Generate client-side class with proxy-class extension methods for OData protocol versions 1.0-3.0 (with generating functions from FunctionImports);
 - Add necessary nuget-packages for client-side OData proxy-classes.
 
 ### Version 0.1.33:
@@ -25,4 +25,11 @@
 *Features:*
 
 - Added methods in OData V3 extension class for calling service methods returns void value.
+
+### Version 0.3:
+
+*Features:*
+
+- Added methods to generating functions from FunctionImports in OData V3 extension class.
+